+++ client/client.go
@@ -12,14 +12,18 @@ import (
 	"regexp"
 	"strconv"
 	"strings"
+	"sync"
 	"sync/atomic"
 	"time"
+
+	"github.com/getlantern/flashlight/config"
+	"github.com/getlantern/flashlight/geolookup"
 
 	lru "github.com/hashicorp/golang-lru"
 
 	"github.com/getlantern/detour"
 	"github.com/getlantern/errors"
-	"github.com/getlantern/eventual"
+	eventual "github.com/getlantern/eventual/v2"
 	"github.com/getlantern/go-socks5"
 	"github.com/getlantern/golog"
 	"github.com/getlantern/hidden"
@@ -53,8 +57,6 @@ var (
 		80, 443,
 		// SSH for those who know how to configure it
 		22,
-		// SMTP and encrypted SMTP
-		25, 465,
 		// POP and encrypted POP
 		110, 995,
 		// IMAP and encrypted IMAP
@@ -117,7 +119,8 @@ type Client struct {
 
 	proxy proxy.Proxy
 
-	l net.Listener
+	httpListener  eventual.Value
+	socksListener eventual.Value
 
 	disconnected         func() bool
 	allowProbes          func() bool
@@ -143,7 +146,17 @@ type Client struct {
 	httpProxyIP   string
 	httpProxyPort string
 
-	chPingProxiesConf chan pingProxiesConf
+	chPingProxiesConf    chan pingProxiesConf
+	googleAdsFilter      func() bool
+	googleAdsOptionsLock sync.RWMutex
+	googleAdsOptions     *config.GoogleSearchAdsOptions
+	adTrackUrl           func() string
+	allowGoogleSearchAds func() bool
+	allowMITM            func() bool
+	eventWithLabel       func(category, action, label string)
+
+	httpWg  sync.WaitGroup
+	socksWg sync.WaitGroup
 }
 
 // NewClient creates a new client that does things like starts the HTTP and
@@ -159,12 +172,15 @@ func NewClient(
 	useDetour func() bool,
 	allowHTTPSEverywhere func() bool,
 	allowMITM func() bool,
+	allowGoogleSearchAds func() bool,
 	userConfig common.UserConfig,
 	statsTracker stats.Tracker,
 	allowPrivateHosts func() bool,
 	lang func() string,
 	adSwapTargetURL func() string,
 	reverseDNS func(addr string) (string, error),
+	adTrackUrl func() string,
+	eventWithLabel func(category, action, label string),
 ) (*Client, error) {
 	// A small LRU to detect redirect loop
 	rewriteLRU, err := lru.New(100)
@@ -189,18 +205,91 @@ func NewClient(
 		allowPrivateHosts:    allowPrivateHosts,
 		lang:                 lang,
 		adSwapTargetURL:      adSwapTargetURL,
+		googleAdsFilter:      allowGoogleSearchAds,
 		reverseDNS:           reverseDNS,
 		chPingProxiesConf:    make(chan pingProxiesConf, 1),
+		googleAdsOptions:     nil,
+		googleAdsOptionsLock: sync.RWMutex{},
+		adTrackUrl:           adTrackUrl,
+		allowGoogleSearchAds: allowGoogleSearchAds,
+		allowMITM:            allowMITM,
+		eventWithLabel:       eventWithLabel,
+		httpListener:         eventual.NewValue(),
+		socksListener:        eventual.NewValue(),
 	}
 
 	keepAliveIdleTimeout := chained.IdleTimeout - 5*time.Second
 
-	var mitmOpts *mitm.Opts
-	if allowMITM() {
+	var mitmErr error
+	client.proxy, mitmErr = proxy.New(&proxy.Opts{
+		IdleTimeout:  keepAliveIdleTimeout,
+		BufferSource: buffers.Pool,
+		Filter:       filters.FilterFunc(client.filter),
+		OnError:      errorResponse,
+		Dial:         client.dial,
+		MITMOpts:     client.MITMOptions(),
+		ShouldMITM: func(req *http.Request, upstreamAddr string) bool {
+			userAgent := req.Header.Get("User-Agent")
+			// Only MITM certain browsers
+			// See http://useragentstring.com/pages/useragentstring.php
+			shouldMITM := strings.Contains(userAgent, "Chrome/") || // Chrome
+				strings.Contains(userAgent, "Firefox/") || // Firefox
+				strings.Contains(userAgent, "MSIE") || strings.Contains(userAgent, "Trident") || // Internet Explorer
+				strings.Contains(userAgent, "Edge") || // Microsoft Edge
+				strings.Contains(userAgent, "QQBrowser") || // QQ
+				strings.Contains(userAgent, "360Browser") || strings.Contains(userAgent, "360SE") || strings.Contains(userAgent, "360EE") // 360
+			return shouldMITM
+		},
+	})
+	if mitmErr != nil {
+		log.Errorf("Unable to initialize MITM: %v", mitmErr)
+	}
+	client.reportProxyLocationLoop()
+	client.iptool, _ = iptool.New()
+	go client.pingProxiesLoop()
+	go func() {
+		for {
+			if geolookup.GetCountry(0) != "" {
+				if err := client.proxy.ApplyMITMOptions(client.MITMOptions()); err != nil {
+					log.Errorf("Unable to initialize MITM: %v", err)
+				}
+				return
+			}
+			<-geolookup.OnRefresh()
+		}
+	}()
+	return client, nil
+}
+
+func (c *Client) MITMOptions() *mitm.Opts {
+	if c.allowMITM() {
 		log.Debug("Enabling MITM")
-		mitmOpts = &mitm.Opts{
-			PKFile:             filepath.Join(configDir, "mitmkey.pem"),
-			CertFile:           filepath.Join(configDir, "mitmcert.pem"),
+
+		domains := []string{
+			// Currently don't bother MITM'ing ad sites since we're not doing ad swapping
+			// "*.doubleclick.net",
+			// "*.g.doubleclick.net",
+			// "adservice.google.com",
+			// "adservice.google.com.hk",
+			// "adservice.google.co.jp",
+			// "adservice.google.nl",
+			// "*.googlesyndication.com",
+			// "*.googletagservices.com",
+			// "googleadservices.com",
+		}
+		// MITM YouTube domains to track statistics on watched videos (list obtained by running ../cmd/youtubescanner/youtubescanner.go)
+		for _, suffix := range MITMSuffixes {
+			domains = append(domains, "*.youtube."+suffix)
+		}
+		// MITM Google search domains to strip the ads/inject relevant data
+		if c.allowGoogleSearchAds() {
+			for _, suffix := range MITMSuffixes {
+				domains = append(domains, "*.google."+suffix)
+			}
+		}
+		return &mitm.Opts{
+			PKFile:             filepath.Join(c.configDir, "mitmkey.pem"),
+			CertFile:           filepath.Join(c.configDir, "mitmcert.pem"),
 			Organization:       "Lantern",
 			InstallCert:        true,
 			InstallPrompt:      i18n.T("BACKEND_MITM_INSTALL_CERT", i18n.T(translationAppName)),
@@ -214,195 +303,10 @@ func NewClient(
 				}
 				op.End()
 			},
-			Domains: []string{
-				// Currently don't bother MITM'ing ad sites since we're not doing ad swapping
-				// "*.doubleclick.net",
-				// "*.g.doubleclick.net",
-				// "adservice.google.com",
-				// "adservice.google.com.hk",
-				// "adservice.google.co.jp",
-				// "adservice.google.nl",
-				// "*.googlesyndication.com",
-				// "*.googletagservices.com",
-				// "googleadservices.com",
-				// MITM YouTube domains to track statistics on watched videos (list obtained by running ../cmd/youtubescanner/youtubescanner.go)
-				"*.youtube.ae",
-				"*.youtube.co.ae",
-				"*.youtube.com.ar",
-				"*.youtube.at",
-				"*.youtube.co.at",
-				"*.youtube.com.au",
-				"*.youtube.az",
-				"*.youtube.com.az",
-				"*.youtube.ba",
-				"*.youtube.be",
-				"*.youtube.bg",
-				"*.youtube.bh",
-				"*.youtube.com.bh",
-				"*.youtube.bo",
-				"*.youtube.com.bo",
-				"*.youtube.com.br",
-				"*.youtube.com.by",
-				"*.youtube.by",
-				"*.youtube.ca",
-				"*.youtube.cat",
-				"*.youtube.ch",
-				"*.youtube.cl",
-				"*.youtube.co",
-				"*.youtube.com.co",
-				"*.youtube.com",
-				"*.youtube.cr",
-				"*.youtube.co.cr",
-				"*.youtube.cz",
-				"*.youtube.de",
-				"*.youtube.dk",
-				"*.youtube.com.do",
-				"*.youtube.com.ec",
-				"*.youtube.ee",
-				"*.youtube.com.ee",
-				"*.youtube.com.eg",
-				"*.youtube.es",
-				"*.youtube.com.es",
-				"*.youtube.fi",
-				"*.youtube.fr",
-				"*.youtube.ge",
-				"*.youtube.com.gh",
-				"*.youtube.gr",
-				"*.youtube.com.gr",
-				"*.youtube.gt",
-				"*.youtube.com.gt",
-				"*.youtube.hk",
-				"*.youtube.com.hk",
-				"*.youtube.com.hn",
-				"*.youtube.hr",
-				"*.youtube.com.hr",
-				"*.youtube.hu",
-				"*.youtube.co.hu",
-				"*.youtube.co.id",
-				"*.youtube.ie",
-				"*.youtube.co.il",
-				"*.youtube.in",
-				"*.youtube.co.in",
-				"*.youtube.iq",
-				"*.youtube.is",
-				"*.youtube.it",
-				"*.youtube.com.jm",
-				"*.youtube.com.jo",
-				"*.youtube.jo",
-				"*.youtube.jp",
-				"*.youtube.co.jp",
-				"*.youtube.co.ke",
-				"*.youtube.kr",
-				"*.youtube.co.kr",
-				"*.youtube.com.kw",
-				"*.youtube.kz",
-				"*.youtube.com.lb",
-				"*.youtube.lk",
-				"*.youtube.lt",
-				"*.youtube.lu",
-				"*.youtube.lv",
-				"*.youtube.com.lv",
-				"*.youtube.ly",
-				"*.youtube.com.ly",
-				"*.youtube.co.ma",
-				"*.youtube.ma",
-				"*.youtube.me",
-				"*.youtube.com.mk",
-				"*.youtube.mk",
-				"*.youtube.com.mt",
-				"*.youtube.mx",
-				"*.youtube.com.mx",
-				"*.youtube.my",
-				"*.youtube.com.my",
-				"*.youtube.ng",
-				"*.youtube.com.ng",
-				"*.youtube.ni",
-				"*.youtube.com.ni",
-				"*.youtube.nl",
-				"*.youtube.no",
-				"*.youtube.co.nz",
-				"*.youtube.com.om",
-				"*.youtube.pa",
-				"*.youtube.com.pa",
-				"*.youtube.pe",
-				"*.youtube.com.pe",
-				"*.youtube.ph",
-				"*.youtube.com.ph",
-				"*.youtube.pk",
-				"*.youtube.com.pk",
-				"*.youtube.pl",
-				"*.youtube.pr",
-				"*.youtube.pt",
-				"*.youtube.com.pt",
-				"*.youtube.qa",
-				"*.youtube.com.py",
-				"*.youtube.com.qa",
-				"*.youtube.ro",
-				"*.youtube.com.ro",
-				"*.youtube.rs",
-				"*.youtube.ru",
-				"*.youtube.sa",
-				"*.youtube.com.sa",
-				"*.youtube.se",
-				"*.youtube.sg",
-				"*.youtube.com.sg",
-				"*.youtube.si",
-				"*.youtube.sk",
-				"*.youtube.sn",
-				"*.youtube.sv",
-				"*.youtube.com.sv",
-				"*.youtube.co.th",
-				"*.youtube.tn",
-				"*.youtube.com.tn",
-				"*.youtube.com.tr",
-				"*.youtube.com.tw",
-				"*.youtube.co.tz",
-				"*.youtube.ua",
-				"*.youtube.com.ua",
-				"*.youtube.ug",
-				"*.youtube.co.ug",
-				"*.youtube.co.uk",
-				"*.youtube.com.uy",
-				"*.youtube.uy",
-				"*.youtube.com.ve",
-				"*.youtube.co.ve",
-				"*.youtube.vn",
-				"*.youtube.voto",
-				"*.youtube.co.za",
-				"*.youtube.co.zw",
-			},
+			Domains: domains,
 		}
 	}
-	var mitmErr error
-	client.proxy, mitmErr = proxy.New(&proxy.Opts{
-		IdleTimeout:  keepAliveIdleTimeout,
-		BufferSource: buffers.Pool,
-		Filter:       filters.FilterFunc(client.filter),
-		OnError:      errorResponse,
-		Dial:         client.dial,
-		MITMOpts:     mitmOpts,
-		ShouldMITM: func(req *http.Request, upstreamAddr string) bool {
-			userAgent := req.Header.Get("User-Agent")
-			// Only MITM certain browsers
-			// See http://useragentstring.com/pages/useragentstring.php
-			shouldMITM := strings.Contains(userAgent, "Chrome/") || // Chrome
-				strings.Contains(userAgent, "MSIE") || strings.Contains(userAgent, "Trident") || // Internet Explorer
-				strings.Contains(userAgent, "Edge") || // Microsoft Edge
-				strings.Contains(userAgent, "QQBrowser") || // QQ
-				strings.Contains(userAgent, "360Browser") || strings.Contains(userAgent, "360SE") || strings.Contains(userAgent, "360EE") // 360
-			return shouldMITM
-		},
-	})
-	if mitmErr != nil {
-		log.Errorf("Unable to initialize MITM: %v", mitmErr)
-	}
-	client.reportProxyLocationLoop()
-	client.iptool, err = iptool.New()
-	if err != nil {
-		return nil, errors.New("Unable to initialize iptool: %v", err)
-	}
-	go client.pingProxiesLoop()
-	return client, nil
+	return nil
 }
 
 func (client *Client) GetBalancer() *balancer.Balancer {
@@ -431,7 +335,10 @@ func (client *Client) reportProxyLocationLoop() {
 // Addr returns the address at which the client is listening with HTTP, blocking
 // until the given timeout for an address to become available.
 func Addr(timeout time.Duration) (interface{}, bool) {
-	return addr.Get(timeout)
+	ctx, cancel := context.WithTimeout(context.Background(), timeout)
+	defer cancel()
+	result, err := addr.Get(ctx)
+	return result, err == nil
 }
 
 // Addr returns the address at which the client is listening with HTTP, blocking
@@ -443,7 +350,10 @@ func (client *Client) Addr(timeout time.Duration) (interface{}, bool) {
 // Socks5Addr returns the address at which the client is listening with SOCKS5,
 // blocking until the given timeout for an address to become available.
 func Socks5Addr(timeout time.Duration) (interface{}, bool) {
-	return socksAddr.Get(timeout)
+	ctx, cancel := context.WithTimeout(context.Background(), timeout)
+	defer cancel()
+	result, err := socksAddr.Get(ctx)
+	return result, err == nil
 }
 
 // Socks5Addr returns the address at which the client is listening with SOCKS5,
@@ -459,6 +369,9 @@ func (client *Client) Socks5Addr(timeout time.Duration) (interface{}, bool) {
 // Sometimes on Windows, http.Server may fail to accept new connections after
 // running for a random period. This method will try serve again.
 func (client *Client) ListenAndServeHTTP(requestedAddr string, onListeningFn func()) error {
+	client.httpWg.Add(1)
+	defer client.httpWg.Done()
+
 	var err error
 	var l net.Listener
 	log.Debugf("About to listen at '%s'", requestedAddr)
@@ -470,8 +383,13 @@ func (client *Client) ListenAndServeHTTP(requestedAddr string, onListeningFn fun
 			return fmt.Errorf("Unable to listen: %q", err)
 		}
 	}
+	defer l.Close()
 
-	client.l = l
+	client.httpListener.Set(l)
+	defer func() {
+		client.httpListener.Reset()
+	}()
+
 	listenAddr := l.Addr().String()
 	addr.Set(listenAddr)
 	client.httpProxyIP, client.httpProxyPort, _ = net.SplitHostPort(listenAddr)
@@ -496,12 +414,22 @@ func (client *Client) ListenAndServeHTTP(requestedAddr string, onListeningFn fun
 // ListenAndServeSOCKS5 starts the SOCKS server listening at the specified
 // address.
 func (client *Client) ListenAndServeSOCKS5(requestedAddr string) error {
+	client.socksWg.Add(1)
+	defer client.socksWg.Done()
+
 	var err error
 	var l net.Listener
 	if l, err = net.Listen("tcp", requestedAddr); err != nil {
 		return fmt.Errorf("Unable to listen: %q", err)
 	}
-	l = &optimisticListener{l}
+	l = &optimisticListener{Listener: l}
+	defer l.Close()
+
+	client.socksListener.Set(l)
+	defer func() {
+		client.socksListener.Reset()
+	}()
+
 	listenAddr := l.Addr().String()
 	socksAddr.Set(listenAddr)
 
@@ -549,7 +477,27 @@ func (client *Client) Configure(proxies map[string]*chained.ChainedServerInfo) [
 // Stop is called when the client is no longer needed. It closes the
 // client listener and underlying dialer connection pool
 func (client *Client) Stop() error {
-	return client.l.Close()
+	httpListener, _ := client.httpListener.Get(eventual.DontWait)
+	socksListener, _ := client.socksListener.Get(eventual.DontWait)
+
+	var httpError error
+	var socksError error
+
+	if httpListener != nil {
+		httpError = httpListener.(net.Listener).Close()
+		client.httpWg.Wait()
+		addr.Reset()
+	}
+	if socksListener != nil {
+		socksError = socksListener.(net.Listener).Close()
+		client.socksWg.Wait()
+		socksAddr.Reset()
+	}
+
+	if httpError != nil {
+		return httpError
+	}
+	return socksError
 }
 
 func (client *Client) dial(ctx context.Context, isConnect bool, network, addr string) (conn net.Conn, err error) {
@@ -754,6 +702,9 @@ func (client *Client) isAddressProxyable(addr string) error {
 			if strings.HasSuffix(host, ".local") {
 				return fmt.Errorf("%v ends in .local, considering private", host)
 			}
+			if strings.HasSuffix(host, ".onion") {
+				return fmt.Errorf("%v ends in .onion, considering private", host)
+			}
 		}
 		// assuming non-private
 		return nil
@@ -811,4 +762,10 @@ func errorResponse(ctx filters.Context, req *http.Request, read bool, err error)
 		Body:       ioutil.NopCloser(bytes.NewBuffer(htmlerr)),
 		StatusCode: http.StatusServiceUnavailable,
 	}
+}
+
+func (client *Client) ConfigureGoogleAds(opts config.GoogleSearchAdsOptions) {
+	client.googleAdsOptionsLock.Lock()
+	client.googleAdsOptions = &opts
+	client.googleAdsOptionsLock.Unlock()
 }