+++ flags.go
@@ -6,7 +6,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/getlantern/flashlight/common"
 	"github.com/jaffee/commandeer"
 	"github.com/mitchellh/mapstructure"
 )
@@ -36,11 +35,15 @@ type Flags struct {
 	ForceConfigCountry string        `flag:"force-config-country" help:"if specified, force config fetches to pretend they're coming from this 2 letter country-code"`
 	ReadableConfig     bool          `flag:"readableconfig" help:"if specified, disables obfuscation of the config yaml so that it remains human readable"`
 	Help               bool          `flag:"help" help:"Get usage help"`
-	NoUiHttpToken      bool          `flag:"noUiHttpToken" help:"don't require a HTTP token from the UI"`
+	NoUiHttpToken      bool          `flag:"no-ui-http-token" help:"don't require a HTTP token from the UI"`
 	Standalone         bool          `flag:"standalone" help:"run Lantern in its own browser window (doesn't rely on system browser)"`
 	Initialize         bool          `flag:"initialize" help:"silently initialize Lantern to a state of having working proxy and exit, typically during installation."`
 	Timeout            time.Duration `flag:"timeout" help:"force stop Lantern with an exit status of -1 after the timeout."`
-	Staging            bool          `flag:"-"`
+	// TODO: Remove this flag when this can be specified directly as a country-selected feature, or
+	// overridden via a config file?
+	ReplicaRustUrl string   `flag:"replica-rust-url" help:"use the replica-rust service at the provided endpoint"`
+	Staging        bool     `flag:"-"`
+	Experiments    []string `flag:"enabled-experiments" help:"comma separated list of experiments to enable"`
 }
 
 func (f Flags) AsMap() map[string]interface{} {
@@ -79,11 +82,7 @@ func ParseFlags() Flags {
 	}
 
 	// here we can define default values
-	cfg := Flags{
-		Country:   "xx",
-		YinbiAddr: common.YinbiServerAddr,
-		AuthAddr:  common.AuthServerAddr,
-	}
+	cfg := Flags{}
 
 	// the following will error on invalid arguments and take env variables starting with LANTERN_ into consideration
 	err := commandeer.LoadArgsEnv(&flagSet{flag.CommandLine}, &cfg, args, "LANTERN_", nil)