+++ ios/config.go
@@ -236,7 +236,6 @@ func (cf *configurer) configureFronting(global *config.Global, timeout time.Dura
 	}
 
 	fronted.Configure(certs, global.Client.FrontedProviders(), "cloudfront", cf.fullPathTo("masquerade_cache"))
-	chained.ConfigureFronting(certs, global.Client.FrontedProviders(), cf.configFolderPath)
 	rt, ok := fronted.NewDirect(timeout)
 	if !ok {
 		return errors.New("Timed out waiting for fronting to finish configuring")