+++ ui/defaults.go
@@ -1,5 +1,0 @@
-// +build !disableresourcerandomization
-
-package ui
-
-var defaultUIAddresses = []string{"localhost:0", "127.0.0.1:0"}